Render additional PNGs; Add shading-less textures
`just generate-keycaps` now additionally outputs PNGs renders of the
SVGs it outputs, and outputs a "texture" version, with the only
difference being that it does not include the simulated keycap side
shading.

It now also avoids the use of css variable for applying the keycap
background colors. This was mainly done to improve compatibility with
some SVG renderers (it's very questionable if this is necessary, since
the SVG either way uses other features which are unsupported by them).