Add command to update icon colors; add `just sync``
Added the new command `just update-icon-palettes`, which updates all of
the color palette defs in a given set of icons to match the colors in
the standard theme (which is the only theme currently). Any elements for
outdated colors that are no longer in the theme are left though
unfortunately, since it's not possible to tell them appart from genuine
defs added while editing the file (like patterns or symbols).

Also added the command `just sync` which is a wrapper around all of the
just commands which edit files, with the intent being that you make sure
to run it before it you commit (I ran it for this commit, hence all of
the changed SVGs).

Updating the colors required parsing files using tree-sitter, and only
updating the byte ranges which needed to be, since I didn't want any
unrelated elements to change in the svg files, which would be the case
if they were deserialized into python objects, edited, and then
serialzed again. This way the diff only shows the elements which were
actually modified/inserted. Only 'update_icon_palettes.py' uses the
tree-sitter parser, and I don't think that is going to change.