Work on icons, focus on modifiers/layers
Worked on a miscalenous selection of key icons, with many changes to
modifier and layer switch keys. A theme was to make the bottom row layer
switches more cohesive in alignment and other aspects, one of the
decisions made being that the icons on the modifier keys should be
aligned to the same edge as the text. This is the same alignment used
by apple uses for their keyboard modifier keys, and I think it looks
better.

More precisely, the following changes were made:
- Added some small hearts to 'cat_face'. I'm much more happy with this
  icon now. :3
- Simplified 'del', by changing the dissapeared cat figure to instead
  just be a blob with cat ears.
- Removed the background and border from the corner letters of
  'n_with_ö', 'e_with_ä', and 'o_with_å'. This both makes them more
  minimal and consistent with the yellow media modifier corners.
- Right aligned the function definition in 'fn' and changed its key icon
  to match the one added to 'umlaut'.
- Left aligned the goober in 'hyper'.
- Added an option icon to 'lalt'.
- Flipped 'super' to make it consistent with the other modifier keys,
  and moved the paw prints to instead go over 'space' and 'umlaut'.
- Added a keypad icon to 'num'.
- Changed back the moon in 'sleep' to use the main accent colors, and
  removed the outline.
- Created the icon 'u_with_umlaut' for the U key, with "¨" in the
  corner.
- Created the icon 'p_with_acute' for the P key, with "`" in the corner.
- Added an icon to 'umlaut' to signify that it activates the purplish
  blue symbol in the corner of 'umlaut's corresponding keys.