Add support for color remapping; Add extra keys and 'rshift'
Added support for specifying additional keys which won't show up in the
3D render, and for specifying color remappings for individual keys in
the layout.

I then also added some extra keys, so that I have the possibility to
replace some keys later after I've ordered them. Notably, this included
adding a '~' with the colors remapped to match the bg_main background.

Lastly I also added the 1u 'rshift' key. It isn't used in the layout,
but I included it in the extra keys in case I want to replace the maths
key later.