Add sample layout; Fix 'gaming' icon border
So before I make a full order on the entire layout I first want to order
a small sample of keycaps. This layout is that sample.
Also took the opportunity to fix the border in the 'gaming' icon. It was
only visible in the non-clipped texture variant, but it bugged me.