Make texture.svg compatible with Inkscape; Generate PDF
Added a step to normalize texture.svg by removing a bunch of newer
SVG features which non-browser renderers don't support. The things that
are normalized are:
- Remove the `use` elements which copy the icons with the shading filter
  applied.
- Remove all hidden elements.
- Replace all pallette color references with their literal colors, and
  change all colors to their corresponding srgb hex codes.
- Remove the palette color elements.
- Remove the embedded fonts.
- Set the fill color in every keycap background rectangle's attributes
  that would previously have been set by CSS rules.
- Remove said CSS rules.
- Remove all uses of `transform-origin`, and calculate the corresponding
  new transform to correct for this. (More on this below.)
With this Inkscape is now able to render the image correctly! With this
capability we now use it generate a PDF of 'texture.svg'.

To be able to calculate the transform-origin offsets I had to write a
SVG path parsing and rendering library capable of calculating their
bounding boxes. I probably spent way too much time on this but it was
really fun. ^^

I decided to not apply these normalization changes to 'preview.svg', as
it's still useful to be able to open that file in BoxySVG with all of
the metadata still there.

Some other misc changes:
- Moved the `svg` module into a folder, i.e. renamed:
  'src/bil/svg.py' -> 'src/lib/svg/__init__.py'
- Moved a bunch of SVG utilities for `svg_builder` into `svg`.
- Prefixed all ids that start with a digit with '_'. Apparently IDs are
  not allowed to start with digits in XML.

*There are definitely additional bigger changes that I can't be bothered
to remember right now, just so you know ;)*