Add support for icon margin; Move scale config to layout
Added support for specifying icon margins in layout configurations,
an additional area around what is defined as the 1u square. This is
required since manufacturers may have templates where the graphics
are meant to extend in this way, so that when the printed sheet has
geometry to fold around the keycap (this is my theory at least).

The spacing of the keys in the Blender scene is kept the same (with no
margin), meaning that they don't change in appearance when changing the
margin.

If you're using this you want to edit your icon SVGs so that their
viewBox's extend the same margin. For instance, for a 10 px margin,
you'd set the viewBox to be "-10 -10 120 120". I only configured the
'sample' layout for now, and therefore only edited the 'esc', 'gaming',
and 'n_with_ö_+' icons.

Other changes:
- Moved the scale configuration to the layout file.
- Updated just create-icon to support specifying margins.