Optimize rendering to allow high texture.png resolutions
This commit optimizes the rendering process of texture.png to allow
for very large scaling factors. Currently I'm able to reliably (?)
render out the sample layout with texture scale set to 30, which results
in a final resolution of 15480x5400 (with `"iconMargin"` set to 32)!

This was done by separating the image into a number of tiles at most
3000 by 3000 in size, which are rendered separately, and finanally
stitched together using Image Magick.

This meant adding a new config option `"textureScale"` to the layout
JSON format, which only applies to texture.png since it's the only one
which supports this new tiling method. I decided not to increase the
scale of the sample layout in this commit, since I want to avoid having
to commit the giant PNGs.

Some highlights in the python library:
- Made the `Vec2` class generic supporting both `int`s and `float`s.
- Renamed `Vec2.promote_float` to `Vec2.promote`.
- Replaced all instances of the the `Theme` class with the new `Config`
  class, and changed it so that all fields in `Theme` come from the
  theme JSON file.
- Renamed the `theme.py` module to `config.py` to better fit with the
  above change.
- Removed texture.pdf from the output.