Increase remaining icon margins for 32x32mm
Changes all icons to have padding around them so that their total size
corresponds to 32x32mm. The three icons used by the sample layout were
changed by 0600bfb, this commits updates the remaining.

Also, I updated the iconMargin to 34 in the sample layout since I think
that it was incorrectly set to 32 in that previous commit.