Qualify new ElementTree type argument
It seems like the class `ElementTree` was updated to include a type
argument, which now defaults to `Element | None`, which represents the
root element. This breaks this codebase, which this commit fixes.

Not sure how this happened, as I haven't updated any inputs in my flake
recent (I don't think) and I've configured my VSCode editor to use the
venv generated by Poetry. But my linter has started to give me errors
anyway.