Generate texture.png using `resvg`
Switches to rendering the texture.png using the `resvg` rendering
library. This was done as a part of an effort to be able to generate
massive svg images. The hope is that `resvg` is more deterministic than
Chromium with Playwright taking a screenshoot, which fails unpredictably
at higher resolutions.

This required the new `render` module, which abstracts the notion of
rendering a large image as a set of individual tiles.

I should probably remove the tiling functionality from browser, as it is
now unused, but I didn't feel like it. :3

Now it's possible to reliably create texture.png rendering with scaling
factors over 10x. I won't commit a version with that factor configured
since the generated file would require so much space.