Fix blender coordinates when texture is scaled
Fixes how the texture coordinates in the blender scenes are calculated,
so that they don't break when texture.png has been scaled. This went
unnoticed for quite some time since the moonlander-mk1 layout hasn't
scaled texture.png. That changed in the previous commit which made it so
that the scale attribute applies to all of the output PNGs, meaning that
texture.png ended up being scaled by x3.