Add `just generate-archive`
Adds support for a new `just generate-archive` command, which generates
one image per key in an layout, and bundles them together in a ZIP
archive.

This was implemented as the new python entry point 'package_archive.py'.
I decided to not have the archive be committed, meaning that you have to
call the command whenever you need it.

This resulted in some pretty big framework refactoring to allow for
better configuring of the layout when building the keyboard SVG.